Fix build on v8 7.6; remove MarkIndependent call
Previously object handles were being marked as independent using the
PersistentBase::MarkIndependent api, which was deprecated in V8 7.4 and
removed in V8 7.6. According to the chromium commit that removed it
(4214933c [0]), all handles are treated as independent, making this call
unnecessary.

After this commit is applied, node-capnp will build against the headers
for V8 7.6.303.22 (for electron 6.1.12).

[0]: https://chromium.googlesource.com/v8/v8/+/4214933c6bb4d1815f07a40eaf008cb1e37ef217