Resolve runtime errors in Electron environment
This commit modifies the initialisation of the capnp module such that it
can work in our electron + typescript + webpack environment:

* The dynamic path lookup for the compiled native module is replaced
with a hard-coded string, because otherwise webpack chokes and fails to
link to the native module correctly.
* The dynamic require hook is disabled, as this makes webpack angry and
cannot be used in a typescript environment anyhow.
* The `import` method is renamed to `importFile`, as `import` is a
typescript keyword and cannot therefore be a method name.
* The build step is modified to invoke node-gyp with the correct target
and headers for compilation against the current electron we're using.