Implement darwin cross-compile tooling for node-capnp
This commit introduces tooling to allow a linux host to compile
node-capnp for a darwin target. To do this, tooling is also provided to
allow for cross-compiling capnproto itself, as this didn't really seem
to function before. It's all still a little bit hacky and relies upon
you having the osxcross toolchain available, but I have at least a
moderate belief that it might work.

Approved-by: Henry Harrod