Improve native module require logic for Webpack integration
Previously, we just if-false'd out the dynamic require logic in order to
make webpack happy and get something to build. This was a bit kludgy
(and deviated from upstream a bit much for my liking), and made things
rather difficult to produce a "cross-compiled" electron build with a
specific native module selected.

This commit changes the logic to test for the presence of a specific
define flag, and if present just require a placeholder path for the native
module. This permits webpack to define this flag and then intercept the require,
patching it with the appropriate resource for the target system. If the
flag is not defined (e.g. on a non-webpacked build), the old dynamic
load code proceeds unhindered.