Add Win32 support to capnp.cc
This commit crudely edits capnp.cc to support compiling for Windows
as well as Linux, mostly using `#ifdef _WIN32` preprocessor guards.

Thankfully, the underlying libraries (`kj` and `libuv`) are already
platform-agnostic, so the changes could be quite superficial.

The code passes the compiler, but due to linker woes, it hasn't yet
been tested.