Support Linux-to-Win32 cross-compilation, using a prebuilt binary
This commit stores a precompiled Windows binary in
`prebuild/win-x64/capnp.node`, and enables cross-compilation from Linux
by simply copying that binary to the output. (True cross-compilation
isn't possible, because `node-gyp` doesn't support it.)

In the interest of avoiding complication in `git clone`, we store the
binary directly in the repo, rather than using a solution like Git LFS
or an external web host. Each time the binary changes (which should be
infrequent), it will increase the amount of data downloaded in a
`git clone` by about 1.5 MiB.