Add automation of Win32 capnp.node builds
This commit adds a GitHub Actions workflow which can only be run
explicitly. It sets up a native Win32 build environment on a Windows
runner, uses it to produce a prebuilt `capnp.node` for Win32, and then
uploads the result as an artifact.

The purpose of this workflow is to provide a prebuild `capnp.node` for
Linux-to-Win32 cross-compilation, since `node-gyp` doesn't support true
cross-compilation from scratch.