Do not freeze Capability objects when they're constructed (#6)
Previously, attempting to pass an RPC interface or an
explicitly-constructed `Capability` as the argument to an RPC call would
produce `TypeError: Cannot assign to read only property 'closed' of
object '#<Capability>'`. The same error does not occur when an RPC
capability is used as a bootstrap interface.

The error is caused by a call to `Object.freeze(this)` at the end of the
`Capability` constructor. The `close()` method can then never be called
without crashing, because it always attempts to mutate
`Capability.closed`, which is a frozen property.

It looks like this use-case (passing a local capability as an argument)
isn't exercised by `capnp-test.js`, so it seems likely that this was
just a simple oversight. As far as I can tell, none of `Capability`'s
properties are mutated elsewhere.

This commit changes the `Object.freeze(this)` call to
`Object.seal(this)`, which fixes the error without any apparent
side-effects.