TF-489: Ensure that connected sockets are non-blocking
On non-Linux platforms, it seems it was possible for us to attempt to
connect to a socket which was not non-blocking, which could result in
the entire event loop getting blocked, which would predictably hang the
entire node process; TF-489.

Previously, we were fixing this for Windows systems, but this seems to
be an issue on macOS as well. This commit extends the fix to all
non-Linux platforms.